refactor
separate common functions and consts
refactor tests
