+++ src/common/default-forwarded-headers.ts
@@ -1,0 +1,13 @@
+export const defaultForwardedHeaders = [
+  "accept",
+  "accept-charset",
+  "accept-encoding",
+  "accept-language",
+  "authorization",
+  "cache-control",
+  "content-type",
+  "cookie",
+  "host",
+  "user-agent",
+  "referer",
+]
