+++ src/next-edge/index.test.ts
@@ -1,7 +1,6 @@
 import {
   createApiHandler,
   filterRequestHeaders,
-  guessCookieDomain,
 } from "./index"
 import express from "express"
 import { NextApiRequest, NextApiResponse } from "next"
@@ -283,76 +282,8 @@ describe("NextJS handler", () => {
   })
 })
 
-describe("cookie guesser", () => {
-  test("uses force domain", async () => {
-    expect(
-      guessCookieDomain("https://localhost", {
-        forceCookieDomain: "some-domain",
-      }),
-    ).toEqual("some-domain")
-  })
-
-  test("does not use any guessing domain", async () => {
-    expect(
-      guessCookieDomain("https://localhost", {
-        dontUseTldForCookieDomain: true,
-      }),
-    ).toEqual(undefined)
-  })
-
-  test("is not confused by invalid data", async () => {
-    expect(
-      guessCookieDomain("5qw5tare4g", {
-        dontUseTldForCookieDomain: true,
-      }),
-    ).toEqual(undefined)
-    expect(
-      guessCookieDomain("https://123.123.123.123.123", {
-        dontUseTldForCookieDomain: true,
-      }),
-    ).toEqual(undefined)
-  })
-
-  test("is not confused by IP", async () => {
-    expect(
-      guessCookieDomain("https://123.123.123.123", {
-        dontUseTldForCookieDomain: true,
-      }),
-    ).toEqual(undefined)
-    expect(
-      guessCookieDomain("https://2001:0db8:0000:0000:0000:ff00:0042:8329", {
-        dontUseTldForCookieDomain: true,
-      }),
-    ).toEqual(undefined)
-  })
-
-  test("uses TLD", async () => {
-    expect(guessCookieDomain("https://foo.localhost", {})).toEqual(
-      "foo.localhost",
-    )
-
-    expect(guessCookieDomain("https://foo.localhost:1234", {})).toEqual(
-      "foo.localhost",
-    )
-
-    expect(
-      guessCookieDomain(
-        "https://spark-public.s3.amazonaws.com/dataanalysis/loansData.csv",
-        {},
-      ),
-    ).toEqual("spark-public.s3.amazonaws.com")
-
-    expect(guessCookieDomain("spark-public.s3.amazonaws.com", {})).toEqual(
-      "spark-public.s3.amazonaws.com",
-    )
-
-    expect(guessCookieDomain("https://localhost/123", {})).toEqual("localhost")
-    expect(guessCookieDomain("https://localhost:1234/123", {})).toEqual(
-      "localhost",
-    )
-  })
-
-  test("filters request headers", async () => {
+describe("filterRequestHeaders", () => {
+  test("correctly filters headers", async () => {
     const headers = {
       accept: "application/json",
       filtered: "any",
