+++ src/next-edge/index.ts
@@ -4,28 +4,16 @@ import { IncomingHttpHeaders } from "http"
 import { isText } from "istextorbinary"
 import { NextApiRequest, NextApiResponse } from "next"
 import parse, { splitCookiesString } from "set-cookie-parser"
-import tldjs from "tldjs"
 import { CreateApiHandlerOptions } from "../type/create-api-handler-options"
-import { getBaseUrl } from "../utils/get-base-url"
+import { getBaseUrl } from "../common/get-base-url"
+import { defaultForwardedHeaders } from "../common/default-forwarded-headers"
+import { processLocationHeader } from "../common/process-location-header"
+import { guessCookieDomain } from "../../next-edge"
 
 export function filterRequestHeaders(
   headers: IncomingHttpHeaders,
   forwardAdditionalHeaders?: string[],
 ): Headers {
-  const defaultForwardedHeaders = [
-    "accept",
-    "accept-charset",
-    "accept-encoding",
-    "accept-language",
-    "authorization",
-    "cache-control",
-    "content-type",
-    "cookie",
-    "host",
-    "user-agent",
-    "referer",
-  ]
-
   const filteredHeaders = new Headers()
   Object.entries(headers).forEach(([key, value]) => {
     const isValid =
@@ -37,8 +25,6 @@ export function filterRequestHeaders(
   return filteredHeaders
 }
 
-const encode = (v: string) => v
-
 /**
  * The NextJS API configuration
  */
@@ -46,22 +32,6 @@ export const config = {
   api: {
     bodyParser: false,
   },
-}
-
-function processLocationHeader(locationHeaderValue: string, baseUrl: string) {
-  if (locationHeaderValue.startsWith(baseUrl)) {
-    return locationHeaderValue.replace(baseUrl, "/api/.ory")
-  }
-
-  if (
-    locationHeaderValue.startsWith("/api/kratos/public/") ||
-    locationHeaderValue.startsWith("/self-service/") ||
-    locationHeaderValue.startsWith("/ui/")
-  ) {
-    return "/api/.ory" + locationHeaderValue
-  }
-
-  return locationHeaderValue
 }
 
 function processSetCookieHeader(
@@ -93,7 +63,7 @@ function processSetCookieHeader(
       ...cookie,
       domain,
       secure,
-      encode,
+      encode: (v: string) => v,
     }))
     .map(({ value, name, ...options }) =>
       serialize(name, value, options as CookieSerializeOptions),
@@ -191,29 +161,4 @@ export function createApiHandler(options: CreateApiHandlerOptions) {
 
     res.end()
   }
-}
-
-export function guessCookieDomain(
-  url: string | undefined,
-  options: CreateApiHandlerOptions,
-) {
-  if (!url || options.forceCookieDomain) {
-    return options.forceCookieDomain
-  }
-
-  if (options.dontUseTldForCookieDomain) {
-    return undefined
-  }
-
-  const parsed = tldjs.parse(url || "")
-
-  if (!parsed.isValid || parsed.isIp) {
-    return undefined
-  }
-
-  if (!parsed.domain) {
-    return parsed.hostname
-  }
-
-  return parsed.domain
 }
